Create a core_tasks top-level dir.
Put a few of the backend/core tasks in it.

Puts some of their base classes/mixins into task/ as they are
not themselves tasks, and are intended to be reusable.
