+++ src/python/pants/bin/BUILD
@@ -45,6 +45,7 @@ target(
     'src/python/pants/backend/jvm:plugin',
     'src/python/pants/backend/project_info:plugin',
     'src/python/pants/backend/python:plugin',
+    'src/python/pants/core_tasks',
   ],
 )
