+++ src/python/pants/core_tasks/BUILD
@@ -1,0 +1,10 @@
+# Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+python_library(
+  name='core_tasks',
+  sources=globs('*.py'),
+  dependencies=[
+    'src/python/pants/goal:task_registrar',
+    'src/python/pants/task',
+  ])
