+++ src/python/pants/task/BUILD
@@ -7,12 +7,15 @@ python_library(
   sources = globs('*.py'),
   dependencies = [
     '3rdparty/python/twitter/commons:twitter.common.collections',
+    'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:fingerprint_strategy',
     'src/python/pants/base:worker_pool',
     'src/python/pants/base:workunit',
+    'src/python/pants/build_graph',
     'src/python/pants/cache',
     'src/python/pants/console:stty_utils',
+    'src/python/pants/goal:workspace',
     'src/python/pants/invalidation',
     'src/python/pants/option',
     'src/python/pants/reporting',
