+++ src/python/pants/task/noop_exec_task.py
@@ -21,19 +21,3 @@ class NoopExecTask(Task):
 
   def execute(self):
     pass
-
-
-class NoopCompile(NoopExecTask):
-  """A no-op that provides a product type that can be used to force scheduling."""
-
-  @classmethod
-  def product_types(cls):
-    return ['ran_compile']
-
-
-class NoopTest(NoopExecTask):
-  """A no-op that provides a product type that can be used to force scheduling."""
-
-  @classmethod
-  def product_types(cls):
-    return ['ran_tests']
