+++ tests/python/pants_test/core_tasks/test_what_changed.py
@@ -12,7 +12,6 @@ from pants.backend.codegen.targets.java_thrift_library import JavaThriftLibrary
 from pants.backend.codegen.targets.python_thrift_library import PythonThriftLibrary
 from pants.backend.core.from_target import FromTarget
 from pants.backend.core.targets.resources import Resources
-from pants.backend.core.tasks.what_changed import WhatChanged
 from pants.backend.core.wrapped_globs import Globs, RGlobs
 from pants.backend.jvm.targets.jar_dependency import JarDependency
 from pants.backend.jvm.targets.jar_library import JarLibrary
@@ -21,6 +20,7 @@ from pants.backend.jvm.targets.scala_jar_dependency import ScalaJarDependency
 from pants.backend.jvm.targets.unpacked_jars import UnpackedJars
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.build_graph.build_file_aliases import BuildFileAliases
+from pants.core_tasks.what_changed import WhatChanged
 from pants.goal.workspace import Workspace
 from pants_test.tasks.task_test_base import ConsoleTaskTestBase
